Tidy & Improve Collapsed Columns Behaviour (#2159)
* Tidy & Improve Collapsed Columns Behaviour

* Fix styling

* Add relevant tests

* Fix styling

---------
