+++ src/Traits/WithCollapsingColumns.php
@@ -4,7 +4,7 @@ namespace Rappasoft\LaravelLivewireTables\Traits;
 
 use Rappasoft\LaravelLivewireTables\Traits\Configuration\CollapsingColumnConfiguration;
 use Rappasoft\LaravelLivewireTables\Traits\Helpers\CollapsingColumnHelpers;
-use Rappasoft\LaravelLivewireTables\Traits\Styling\HasCollapsingColumnsStyling;
+use Rappasoft\LaravelLivewireTables\Traits\Styling\Columns\HasCollapsingColumnsStyling;
 
 trait WithCollapsingColumns
 {
@@ -13,4 +13,10 @@ trait WithCollapsingColumns
         HasCollapsingColumnsStyling;
 
     protected bool $collapsingColumnsStatus = true;
+
+    protected ?bool $shouldAlwaysCollapse;
+
+    protected ?bool $shouldMobileCollapse;
+
+    protected ?bool $shouldTabletCollapse;
 }
